+++ LegendWeathers/Utils/TweaksIntegration.cs
@@ -30,7 +30,7 @@ namespace LegendWeathers.Utils
                 RegisterCombinedWeather("Eclipsed + Majora Moon", eclipsed, majoramoon);
                 RegisterCombinedWeather("Majora Chaos", rainy, stormy, eclipsed, majoramoon);
 
-                if (Plugin.config.LethalElementsInstalled)
+                if (Compatibility.LethalElementsInstalled)
                 {
                     RegisterCombinedWeather("Heatwave + Majora Moon", heatwave, majoramoon);
                     RegisterCombinedWeather("Snowfall + Majora Moon", snowfall, majoramoon);
@@ -38,23 +38,23 @@ namespace LegendWeathers.Utils
                     RegisterCombinedWeather("Majora Climate Anomaly", heatwave, solarflare, snowfall, majoramoon);
 
                     string eowName = "The End of the World";
-                    if (Plugin.config.CodeRebirthInstalled && Plugin.config.MrovWeathersInstalled)
+                    if (Compatibility.CodeRebirthInstalled && Compatibility.MrovWeathersInstalled)
                         RegisterCombinedWeather(eowName, rainy, stormy, eclipsed, heatwave, solarflare, snowfall, tornado, blackout, majoramoon);
-                    else if (Plugin.config.CodeRebirthInstalled && !Plugin.config.MrovWeathersInstalled)
+                    else if (Compatibility.CodeRebirthInstalled && !Compatibility.MrovWeathersInstalled)
                         RegisterCombinedWeather(eowName, rainy, stormy, eclipsed, heatwave, solarflare, snowfall, tornado, majoramoon);
-                    else if (!Plugin.config.CodeRebirthInstalled && Plugin.config.MrovWeathersInstalled)
+                    else if (!Compatibility.CodeRebirthInstalled && Compatibility.MrovWeathersInstalled)
                         RegisterCombinedWeather(eowName, rainy, stormy, eclipsed, heatwave, solarflare, snowfall, blackout, majoramoon);
                     else
                         RegisterCombinedWeather(eowName, rainy, stormy, eclipsed, heatwave, solarflare, snowfall, majoramoon);
                 }
 
-                if (Plugin.config.CodeRebirthInstalled)
+                if (Compatibility.CodeRebirthInstalled)
                 {
                     RegisterCombinedWeather("Tornado + Majora Moon", tornado, majoramoon);
                     RegisterCombinedWeather("Majora Superstorm", rainy, stormy, tornado, majoramoon);
                 }
 
-                if (Plugin.config.MrovWeathersInstalled)
+                if (Compatibility.MrovWeathersInstalled)
                 {
                     RegisterCombinedWeather("Blackout + Majora Moon", blackout, majoramoon);
                 }